# perf: Improve import time by lazy import "slow" modules