This PR improves the import time by postponing the import of `asyncio`, `logging`, and `html`. And also not updating signatures if not inside a IPython environment. 

# Import time + tuna
Generated with `python -X importtime -c 'import param' 2> tuna.log && tuna tuna.log`. Some notes only focus on the param and ignore the `param.version` time; also, the measurement is not exact (see hyperfine) and needs warmup time after a branch change. 

Main (https://github.com/holoviz/param/commit/9ff6a9f9c7c36cf2e760f18807d7c538bb42e255):
![Screenshot 2025-02-24 at 12-23-58 tuna - tuna log](https://github.com/user-attachments/assets/f27da201-a2fa-4b4e-a3fd-9846ea188a35)

After removing asyncio (https://github.com/holoviz/param/pull/1037/commits/5a63a9e5f5027f31a10a6f2e7ef40485ec1720b5):
![Screenshot 2025-02-24 at 12-28-03 tuna - tuna log](https://github.com/user-attachments/assets/461bbdba-75c1-4a68-8b38-7fd1ad3a288b)

`+` After removing logging and HTML (https://github.com/holoviz/param/pull/1037/commits/3d4885b4a30b8dfb3689a56421fbf510087e4752)
![image](https://github.com/user-attachments/assets/c4cb3b62-7602-4983-aaa7-5fb3d390da17)

`+` Without updating signature in `__init_submodules__` (https://github.com/holoviz/param/pull/1037/commits/30180b86a716d6f17ec72d7bf3a8516174f04cc5):
![image](https://github.com/user-attachments/assets/1dcafbcf-a73a-4474-9328-6220b38c49dd)
 
# Pyinstrument
Also used pyinstrument with `pyinstrument -r html example.py` where example.py is `import param` 

Main (9ff6a9f)
![image](https://github.com/user-attachments/assets/f343d25c-e830-47cd-aaf3-211a06c5501a)


This PR (https://github.com/holoviz/param/pull/1037/commits/30180b86a716d6f17ec72d7bf3a8516174f04cc5):
![image](https://github.com/user-attachments/assets/2c9807b3-fa19-427b-a9dd-b8695c1aaaa6)
 
# hyperfine

Main (https://github.com/holoviz/param/commit/9ff6a9f9c7c36cf2e760f18807d7c538bb42e255)
``` bash
❯ hyperfine "python -c ''" "python -c 'import param'" --warmup 5
Benchmark 1: python -c ''
  Time (mean ± σ):      15.9 ms ±   0.6 ms    [User: 11.8 ms, System: 3.9 ms]
  Range (min … max):    14.4 ms …  17.6 ms    180 runs

Benchmark 2: python -c 'import param'
  Time (mean ± σ):      57.9 ms ±   1.2 ms    [User: 47.5 ms, System: 9.9 ms]
  Range (min … max):    56.0 ms …  61.2 ms    50 runs

Summary
  python -c '' ran
    3.65 ± 0.16 times faster than python -c 'import param'

```

This PR (https://github.com/holoviz/param/pull/1037/commits/30180b86a716d6f17ec72d7bf3a8516174f04cc5):
``` bash
❯ hyperfine "python -c ''" "python -c 'import param'" --warmup 5
Benchmark 1: python -c ''
  Time (mean ± σ):      15.8 ms ±   0.6 ms    [User: 11.7 ms, System: 4.0 ms]
  Range (min … max):    14.4 ms …  17.6 ms    187 runs

Benchmark 2: python -c 'import param'
  Time (mean ± σ):      37.1 ms ±   0.9 ms    [User: 29.8 ms, System: 7.0 ms]
  Range (min … max):    35.4 ms …  39.2 ms    79 runs

Summary
  python -c '' ran
    2.35 ± 0.11 times faster than python -c 'import param'
```