@hoxbro 2025-02-26
> What's the Python version used for all these benchmarks? 

`Python 3.12.9 | packaged by conda-forge | (main, Feb 14 2025, 08:00:06) [GCC 13.3.0]`

> I recommend opening 2 PRs, one for the deferred stdlib imports, one for the signature.

Honestly, I would rather not. I don't think there is any discussion about deferring imports, only around the signature, and both are related to import time.