@maximlt 2025-02-26
> Honestly, I would rather not. I don't think there is any discussion about deferring imports, only around the signature, and both are related to import time.

Sorry I wasn't clear enough. It was not about discussing the changes, but about separating them. The deferred imports are an implementation detail imo, while the updated signature is a breaking change.