@hoxbro 2025-02-26
> updated signature is a breaking change.

Hopefully not? If it was always meant to only work in IPython it should not be a breaking change.