@hoxbro 2025-02-27
Split into two PRs see https://github.com/holoviz/param/pull/1038.